+++ stm32handC5_411CEU6/Src/time_detection.h
@@ -13,7 +13,7 @@
 #include <stdlib.h>
 
 /* Function Declarations */
-extern int time_detection(const double data[], int data_size, int win_size_time, int
+extern int time_detection(const float data[], int data_size, int win_size_time, int
   stride_time, int time_times, int time_add);
 
 #endif
